functions/requirements.txt
pulsar-client==2.1.0
flake8==3.6.0
wheel==0.32.3